% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_functions.R
\encoding{UTF-8}
\name{cyr_lat}
\alias{cyr_lat}
\title{Serbian Cyrillic to Serbian Latin transliterator}
\usage{
cyr_lat(df)
}
\arguments{
\item{df}{can be a data frame, tibble or character string}
}
\value{
data frame, tibble or character string converted to Serbian Latin characters
}
\description{
Converting Serbian Cyrillic characters to Serbian Latin characters.
The function, changes locale settings of R session to Serbian Latin
for easier manipulation with Serbian Latin character sets.
When this function converts data frame or tibble function does not
change the characters of column names. In the future it is planed to
add this option.
}
\examples{
\dontrun{
lcl(loc = "C")
cyr_lat(c("Ч", "љ", "ђ", "ж", "ћ", "њ"))

lcl(loc = "C")
df_text <- data.frame(num = 1:6, text = c("Ч", "љ", "ђ", "ж", "ћ", "њ"),
stringsAsFactors = FALSE)
cyr_lat(df_text)
}
}
