% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_functions.R
\name{lat_cir}
\alias{lat_cir}
\title{Converts Serbian Latin strings to Serbian Cyrillic strings}
\usage{
lat_cir(df)
}
\arguments{
\item{df}{can be a data frame, tibble or character string}
}
\value{
data frame, tibble or character string converted to Serbian Cyrillic characters
}
\description{
Converts Serbian Latin strings to Serbian Cyrillic strings
}
\examples{
text <- "Čačkanje Ljuljaškice Ćuteći Žestoko Đuradj"
lat_cir(text)

df_text <- data.frame(num = 1:5, text = c("Čačkanje", "Ljuljaškice",
"Ćuteći", "Žestoko", "Đuradj"),
stringsAsFactors = FALSE)
lat_cir(df_text)
}
