% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_functions.R
\name{lat_cyr}
\alias{lat_cyr}
\title{Serbian Latin to Serbian Ciryllic transliterator}
\usage{
lat_cyr(df)
}
\arguments{
\item{df}{can be a data frame, tibble or character string}
}
\value{
data frame, tibble or character string converted to Serbian Cyrillic characters.
}
\description{
Converting Serbian Latin characters to Serbian Cyrillic caracters.
The function, changes locale settings of R session to Serbian Cyrillic
for easier mannipulation with Serbian Cyrillic character sets.
}
\examples{
\dontrun{
lat_cyr(c("Č", "ž", "đ", "lj", "ć", "dž"))

df_text <- data.frame(num = 1:6, c("Č", "ž", "đ", "lj", "ć", "dž"),
stringsAsFactors = FALSE)
lat_cyr(df_text)
}
}
