% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_functions.R
\docType{data}
\name{myenv}
\alias{myenv}
\title{Function for replacing latin to cirilic characters from string
Function for replacing latin to cirilic characters from string}
\format{An object of class \code{environment} of length 2.}
\usage{
myenv
}
\description{
Function for replacing latin to cirilic characters from string
Function for replacing latin to cirilic characters from string
}
\keyword{datasets}
