% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_functions.R
\name{test_lat_cir}
\alias{test_lat_cir}
\title{Function for replacing latin to cirilic characters from string
Function for replacing latin to cirilic characters from string}
\usage{
test_lat_cir(df)
}
\description{
Function for replacing latin to cirilic characters from string
Function for replacing latin to cirilic characters from string
}
