% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_functions.R
\name{cyr_lat}
\alias{cyr_lat}
\title{Converts Serbian Cyrillic strings to Serbian Latin strings}
\usage{
cyr_lat(df)
}
\arguments{
\item{df}{can be a data frame, tibble or character string}
}
\value{
data frame, tibble or character string converted to Serbian Latin characters
}
\description{
Converts Serbian Cyrillic strings to Serbian Latin strings
}
\examples{
text <- "Чачкање Љуљашкице Ћутећи Жестоко Ђурађ"
cyr_lat(text)

df_text <- data.frame(num = 1:5, text = c("Чачкање", "Љуљашкице",
"Ћутећи", "Жестоко", "Ђурађ"),
stringsAsFactors = FALSE)
cyr_lat(df_text)
}
