% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_functions.R
\name{cyr_lat}
\alias{cyr_lat}
\title{Serbian Cyrillic to Serbian Latin transliterator}
\usage{
cyr_lat(df)
}
\arguments{
\item{df}{can be a data frame, tibble or character string}
}
\value{
data frame, tibble or character string converted to Serbian Latin characters
}
\description{
Converting Serbian Cyrillic characters to Serbian Latin caracters.
The function, changes locale settings of R session to Serbian Latin
for easier mannipulation with Serbian Latin character sets.
When this function converts data frame or tibble function does not
change the characters of column names. In the future it is planed to
add this option.
}
\examples{
\dontrun{
cyr_lat(c("Ч", "љ", "ђ", "ж", "ћ", "њ"))

df_text <- data.frame(num = 1:6, c("Ч", "љ", "ђ", "ж", "ћ", "њ"),
stringsAsFactors = FALSE)
cyr_lat(df_text)
}
}
