% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_functions.R
\name{lat_cyr}
\alias{lat_cyr}
\title{Serbian Latin to Serbian Ciryllic transliterator}
\usage{
lat_cyr(df)
}
\arguments{
\item{df}{can be a data frame, tibble or character string}
}
\value{
data frame, tibble or character string converted to Serbian Cyrillic characters.
}
\description{
Converting Serbian Latin characters to Serbian Cyrillic characters.
The function, changes locale settings of R session to Serbian Cyrillic
for easier manipulation with Serbian Cyrillic character sets.
When this function converts data frame or tibble function does not
change the characters of column names. In the future it is planed to
add this option.
}
\examples{
\dontrun{
locale(loc = "L") #Important!
lat_cyr(c("Č", "ž", "đ", "lj", "ć", "dž"))

locale(loc = "L") #Important!
df_text <- data.frame(num = 1:6, c("Č", "ž", "đ", "lj", "ć", "dž"),
stringsAsFactors = FALSE)
lat_cyr(df_text)
}
}
