% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locale.R
\name{locale}
\alias{locale}
\title{Setting locale}
\usage{
locale(loc)
}
\arguments{
\item{loc}{string argument for sys.setlocale(), L-Serbian Latin locale,
C-Serbian Cyrillic locale settings and D-for system default settings.}
}
\description{
Setting locale
}
